added package classNames and created components for form
