+++ package.json
@@ -51,6 +51,7 @@
     "@types/react-redux": "^7.1.25",
     "bootstrap": "^5.2.3",
     "bootstrap-icons": "^1.10.3",
+    "classnames": "^2.3.2",
     "concurrently": "^7.6.0",
     "ky": "^0.33.2",
     "module-alias": "^2.2.2",
