+++ src/components/LoginForm.tsx
@@ -1,15 +1,15 @@
-import React, { FC } from "react";
+import React from "react";
 import { NavLink, useNavigate } from "react-router-dom";
-import { Button, Form, Container, Card } from "react-bootstrap";
+import { Button, Form } from "react-bootstrap";
 import { useForm } from "react-hook-form";
 import { yupResolver } from "@hookform/resolvers/yup";
 import * as Yup from "yup";
 import ky from "ky";
 import { toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
-import "@/styles/Auth.scss";
 import { useDispatch } from "react-redux";
 import { loginUser } from "@/store/actionCreators";
+import classNames from "classnames";
 
 type UserLoginForm = {
   email: string;
@@ -21,9 +21,10 @@ type ResponseLogin = {
   message: string;
 };
 
-const LoginPage: FC = () => {
+const LoginForm = () => {
   const navigate = useNavigate();
   const dispatch = useDispatch();
+
   const validationSchemaLogin = Yup.object().shape({
     email: Yup.string().required("Email обязателен").email("Неверный email"),
     password: Yup.string()
@@ -31,6 +32,7 @@ const LoginPage: FC = () => {
       .min(6, "Пароль должен состьять из 6 символов")
       .max(20, "Пароль не может иметь больше 20 сиволов"),
   });
+
   const {
     register,
     handleSubmit,
@@ -38,6 +40,14 @@ const LoginPage: FC = () => {
   } = useForm<UserLoginForm>({
     resolver: yupResolver(validationSchemaLogin),
   });
+
+  const formClassEmail: string = classNames("form-control", {
+    "is-invalid": errors.email,
+  });
+  const formClassPassword: string = classNames("form-control", {
+    "is-invalid": errors.password,
+  });
+
   const onSubmitLogin: any = async (data: UserLoginForm) => {
     try {
       const response: ResponseLogin = await ky
@@ -55,43 +65,37 @@ const LoginPage: FC = () => {
       }
     }
   };
+
   return (
-    <div>
-      <Container className="d-flex justify-content-center align-items-center ">
-        <Card className="p-5 m-5 FormLogin">
-          <h2 className="m-auto">Авторизация</h2>
-          <Form
-            className="d-flex flex-column mt-3"
-            onSubmit={handleSubmit(onSubmitLogin)}
-          >
-            <Form.Label>Email</Form.Label>
-            <Form.Control
-              type="email"
-              {...register("email")}
-              className={`form-control ${errors.email ? "is-invalid" : ""}`}
-            />
-            <div className="invalid-feedback">{errors.email?.message}</div>
-            <Form.Label>Пароль</Form.Label>
-            <Form.Control
-              type="password"
-              {...register("password")}
-              className={`form-control ${errors.password ? "is-invalid" : ""}`}
-            />
-            <div className="invalid-feedback">{errors.password?.message}</div>
-            <div className=" d-flex  justify-content-between mt-3 pl-3 pr-3">
-              <div>
-                Нет аккаунта?
-                <NavLink to="/registration">Зарегистрироваться</NavLink>
-              </div>
-              <Button variant="outline-light" type="submit">
-                Войти
-              </Button>
-            </div>
-          </Form>
-        </Card>
-      </Container>
-    </div>
+    <Form
+      className="d-flex flex-column mt-3"
+      onSubmit={handleSubmit(onSubmitLogin)}
+    >
+      <Form.Label>Email</Form.Label>
+      <Form.Control
+        type="email"
+        {...register("email")}
+        className={formClassEmail}
+      />
+      <div className="invalid-feedback">{errors.email?.message}</div>
+      <Form.Label>Пароль</Form.Label>
+      <Form.Control
+        type="password"
+        {...register("password")}
+        className={formClassPassword}
+      />
+      <div className="invalid-feedback">{errors.password?.message}</div>
+      <div className=" d-flex  justify-content-between mt-3 pl-3 pr-3">
+        <div>
+          Нет аккаунта?
+          <NavLink to="/registration">Зарегистрироваться</NavLink>
+        </div>
+        <Button variant="outline-light" type="submit">
+          Войти
+        </Button>
+      </div>
+    </Form>
   );
 };
 
-export default LoginPage;
+export default LoginForm;
